+++ res/values/strings.xml
@@ -310,10 +310,9 @@
 
     <!-- Search -->
     <string name="search_menu_item">Search</string>
-    <string name="search_bar_hint">Searching...</string>
+    <string name="search_bar_hint">Searching&#8230;</string>
     <string name="search_txt_empty">Please input keywords</string>
-    <string name="search_content_empty">Can not find the searched data</string>
-    <string name="search_server_not_support">Search not supported on server</string>
-    <string name="search_open_dir_not_support">Sorry, open folder not supported yet!</string>
+    <string name="search_content_empty">Nothing was found</string>
+    <string name="search_server_not_support">Oops, Search was not supported on this server!</string>
     <string name="search_library_not_found">Couldn\'t find this library. It may be deleted</string>
 </resources>
