+++ src/com/seafile/seadroid2/ui/WidgetUtils.java
@@ -1,0 +1,76 @@
+package com.seafile.seadroid2.ui;
+
+import android.content.ActivityNotFoundException;
+import android.content.Context;
+import android.content.DialogInterface;
+import android.content.Intent;
+import android.net.Uri;
+import android.support.v4.app.FragmentActivity;
+import android.webkit.MimeTypeMap;
+import com.seafile.seadroid2.R;
+import com.seafile.seadroid2.ui.activity.BrowserActivity;
+import com.seafile.seadroid2.ui.activity.MarkdownActivity;
+import com.seafile.seadroid2.ui.dialog.OpenAsDialog;
+
+import java.io.File;
+
+/**
+ * Activity Utils
+ */
+public class WidgetUtils {
+
+    public static void showFile(final FragmentActivity activity, File file) {
+
+        String name = file.getName();
+        String suffix = name.substring(name.lastIndexOf('.') + 1).toLowerCase();
+
+        if (suffix.length() == 0) {
+            ToastUtils.show(activity, R.string.unknown_file_type);
+            return;
+        }
+
+        // Open markdown files in MarkdownActivity
+        if (suffix.endsWith("md") || suffix.endsWith("markdown")) {
+            startMarkdownActivity(activity, file.getPath());
+            activity.finish();
+            activity.overridePendingTransition(0, 0);
+            return;
+        }
+
+        String mime = MimeTypeMap.getSingleton().getMimeTypeFromExtension(suffix);
+        Intent open = new Intent(Intent.ACTION_VIEW);
+        open.setDataAndType((Uri.fromFile(file)), mime);
+
+        try {
+            activity.startActivity(open);
+            activity.finish();
+            return;
+        } catch (ActivityNotFoundException e) {
+            new OpenAsDialog(file) {
+                @Override
+                public void onDismiss(DialogInterface dialog) {
+                    activity.finish();
+                }
+            }.show(activity.getSupportFragmentManager(), "OpenAsDialog");
+            return;
+        }
+
+    }
+
+    public static void showRepo(Context context, String repoID, String repoName, String path, String dirID) {
+        Intent intent = new Intent(context, BrowserActivity.class);
+        intent.putExtra("repoID", repoID);
+        intent.putExtra("repoName", repoName);
+        intent.putExtra("path", path);
+        intent.putExtra("dirID", dirID);
+        context.startActivity(intent);
+    }
+
+    public static void startMarkdownActivity(Context context, String path) {
+        Intent intent = new Intent(context, MarkdownActivity.class);
+        intent.putExtra("path", path);
+        context.startActivity(intent);
+    }
+
+
+}
