+++ src/com/seafile/seadroid2/ui/activity/CreateGesturePasswordActivity.java
@@ -19,7 +19,7 @@ import com.seafile.seadroid2.gesturelock.LockPatternUtils;
 import com.seafile.seadroid2.gesturelock.LockPatternView;
 import com.seafile.seadroid2.gesturelock.LockPatternView.Cell;
 import com.seafile.seadroid2.gesturelock.LockPatternView.DisplayMode;
-import com.seafile.seadroid2.util.ToastUtils;
+import com.seafile.seadroid2.ui.ToastUtils;
 
 public class CreateGesturePasswordActivity extends Activity implements
         OnClickListener {
