+++ src/com/seafile/seadroid2/ui/activity/UnlockGesturePasswordActivity.java
@@ -20,7 +20,7 @@ import com.seafile.seadroid2.SettingsManager;
 import com.seafile.seadroid2.gesturelock.LockPatternUtils;
 import com.seafile.seadroid2.gesturelock.LockPatternView;
 import com.seafile.seadroid2.gesturelock.LockPatternView.Cell;
-import com.seafile.seadroid2.util.ToastUtils;
+import com.seafile.seadroid2.ui.ToastUtils;
 
 
 public class UnlockGesturePasswordActivity extends Activity {
